+++ nenupy/astro/astro.py
@@ -20,7 +20,9 @@
     * :func:`~nenupy.astro.astro.to_altaz`: Convert :class:`~astropy.coordinates.ICRS` to :class:`~astropy.coordinates.AltAz`
     * :func:`~nenupy.astro.astro.ho_zenith`: Get the local zenith in :class:`~astropy.coordinates.AltAz` coordinates
     * :func:`~nenupy.astro.astro.eq_zenith`: Get the local zenith in :class:`~astropy.coordinates.ICRS` coordinates
-    * :func:`~nenupy.astro.astro.radio_sources`: Get main radio source poisitons
+    * :func:`~nenupy.astro.astro.radio_sources`: Get main radio source positions
+    * :func:`~nenupy.astro.astro.getSource`: Get a particular source coordinates
+    * :func:`~nenupy.astro.astro.altazProfile`: Retrieve horizontal coordinates versus time
     * :func:`~nenupy.astro.astro.meridian_transit`: Next meridian transit time
     * :func:`~nenupy.astro.astro.dispersion_delay`: Dispersion delay induced by wave propagation through an electron plasma
 
@@ -45,13 +47,17 @@ __all__ = [
     'ho_zenith',
     'eq_zenith',
     'radio_sources',
+    'getSource',
+    'altazProfile',
     'meridian_transit',
-    'dispersion_delay'
+    'dispersion_delay',
+    '_pos2ecef',
+    '_ecef2enu'
     ]
 
 
 import numpy as np
-from astropy.time import Time
+from astropy.time import Time, TimeDelta
 from astropy import units as u
 from astropy.coordinates import (
     EarthLocation,
@@ -64,6 +70,14 @@ from astropy.coordinates import (
     get_body
 )
 from astropy.constants import c as lspeed
+from os.path import dirname, join
+import json
+
+from nenupy.instru import nenufar_loc
+import nenupy.miscellaneous as misc
+
+import logging
+log = logging.getLogger(__name__)
 
 
 # ============================================================= #
@@ -97,6 +111,7 @@ def nenufar_loc():
 # ============================================================= #
 # ---------------------------- lst ---------------------------- #
 # ============================================================= #
+@misc.time_args('time')
 def lst(time):
     """ Local sidereal time
 
@@ -106,10 +121,6 @@ def lst(time):
         :returns: LST time
         :rtype: :class:`~astropy.coordinates.Angle`
     """
-    if not isinstance(time, Time):
-        raise TypeError(
-            'time is not an astropy Time.'
-            )
     location = nenufar_loc()
     lon = location.to_geodetic().lon
     lst = time.sidereal_time('apparent', lon)
@@ -304,11 +315,11 @@ def to_altaz(radec, time):
                 )
     """
     if isinstance(radec, (ICRS, SkyCoord)):
-        if isinstance(radec, SkyCoord):
-            if not isinstance(radec.frame, ICRS):
-                raise TypeError(
-                    'frame should be ICRS'
-                )
+        # if isinstance(radec, SkyCoord):
+        #     if not isinstance(radec.frame, ICRS):
+        #         raise TypeError(
+        #             'frame should be ICRS'
+        #         )
         pass
     else:
         raise TypeError(
@@ -438,8 +449,177 @@ def radio_sources(time):
 
 
 # ============================================================= #
-# ----------------------- radio_sources ----------------------- #
+# ------------------------- getSource ------------------------- #
 # ============================================================= #
+def getSource(name, time=None):
+    """ Retrieve a source equatorial coordinates.
+
+        :param name:
+            Source name.
+        :type name: `str`
+        :param time:
+            Time at which the coordinates need to be calculated.
+            This is particularly relevant for Solar System objects.
+            Moreover, if ``time`` is not a scalar, the returned
+            source coordinates will have dimensions matching the
+            ``time`` argument.
+        :type time: :class:`~astropy.time.Time`
+
+        :returns:
+            Source equatorial coordinates.
+        :rtype: :class:`~astropy.coordinates.SkyCoord`
+
+        :Example:
+            >>> from nenupy.astro import getSource
+            >>> from astropy.time import Time
+            >>> getSource('Cyg X-3')
+                <SkyCoord (ICRS): (ra, dec) in deg
+                [(308.10741667, 40.95775)]>
+            >>> getSource('Sun', Time(['2020-04-01 12:00:00', '2020-04-01 14:00:00']))
+                <SkyCoord (GCRS:
+                    obstime=['2020-04-01 12:00:00.000' '2020-04-01 14:00:00.000'],
+                    obsgeoloc=[(4237729.57730929,  917401.20083485, 4662142.24721379),
+                    (3208307.85440129, 2913433.12950414, 4664141.44068417)] m,
+                    obsgeovel=[( -66.89938865, 308.36222659, 0.13071683),
+                    (-212.45197538, 233.29547849, 0.41179784)] m / s): (ra, dec, distance) in (deg, deg, AU)
+                    [(10.983783  , 4.71984404, 0.99938543),
+                     (11.05889498, 4.75192569, 0.99941305)]>
+
+        .. versionadded:: 1.1.0
+    """
+    sourceJson = join(
+        dirname(__file__),
+        'radio_sources.json',
+    )
+    with open(sourceJson) as jsonFile:
+        sources = json.load(jsonFile)
+
+    solarSystem = [
+        'sun',
+        'moon',
+        'jupiter',
+        'saturn',
+        'mars',
+        'venus',
+        'uranus',
+        'neptune'
+    ]
+
+    nTime = 1
+
+    if name.upper() in sources.keys():
+        if time is not None:
+            if not time.isscalar:
+                nTime = time.size
+            else:
+                nTime = 1
+        src = sources[name.upper()]
+        src = SkyCoord(
+            ra=[src['ra']] * nTime * u.deg,
+            dec=[src['dec']] * nTime * u.deg
+        )
+        log.info(
+            f'Source {name} found in {sourceJson}.'
+        )
+    elif name.lower() in solarSystem:
+        if not isinstance(time, Time):
+            raise TypeError(
+                'time should be a Time object'
+            )
+        with solar_system_ephemeris.set('builtin'):
+            src = get_body(name, time, nenufar_loc())
+        log.info(
+            f'Source {name} found in Solar System Ephemeris.'
+        )
+    else:
+        if time is not None:
+            if not time.isscalar:
+                nTime = time.size
+            else:
+                nTime = 1
+        src = SkyCoord.from_name(name)
+        if time is not None:
+            if not time.isscalar:
+                nTime = time.size
+        else:
+            nTime = 1
+        src = SkyCoord(
+            ra=[src.ra.deg] * nTime * u.deg,
+            dec=[src.dec.deg] * nTime * u.deg
+        )
+        log.info(
+            f'Source {name} found in Simbad.'
+        )
+
+    return src
+# ============================================================= #
+
+
+# ============================================================= #
+# ----------------------- altazProfile ------------------------ #
+# ============================================================= #
+def altazProfile(sourceName, tMin=None, tMax=None, dt=None):
+    """
+    """
+    if tMin is None:
+        tMin = Time.now()
+    elif not isinstance(tMin, Time):
+        raise TypeError(
+            'tMin should be a Time object'
+        )
+
+    if tMax is None:
+        tMax = tMin + TimeDelta(24*3600, format='sec')
+    elif not isinstance(tMax, Time):
+        raise TypeError(
+            'tMax should be a Time object'
+        )
+
+    if dt is None:
+        dt = TimeDelta(5*60, format='sec')
+    elif not isinstance(dt, TimeDelta):
+        raise TypeError(
+            'dt should be a TimeDelta object'
+        )
+
+    tCurrent = tMin
+    times = []
+    elevations = []
+    azimuths = []
+    while tCurrent <= tMax:
+        if 'srcRaDec' not in locals():
+            srcRaDec = getSource(
+                name=sourceName,
+                time=tCurrent
+            )
+        elif srcRaDec.frame.name == 'gcrs':
+            # Solar system source, need to be updated/time
+            srcRaDec = getSource(
+                name=sourceName,
+                time=tCurrent
+            )
+        else:
+            # Do not ask again for same source
+            pass
+
+        srcAltAz = to_altaz(
+            radec=srcRaDec,
+            time=tCurrent
+        )
+
+        times.append(tCurrent.isot)
+        azimuths.append(srcAltAz.az.deg)
+        elevations.append(srcAltAz.alt.deg)
+        tCurrent += dt
+
+    return Time(times), np.array(azimuths), np.array(elevations)
+# ============================================================= #
+
+
+# ============================================================= #
+# --------------------- meridian_transit ---------------------- #
+# ============================================================= #
+@misc.time_args('from_time')
 def meridian_transit(source, from_time, npoints=400):
     """ Find the next ``source``meridian transit time after the
         time ``from_time`` at NenuFAR location. This is a wrapper
@@ -468,10 +648,6 @@ def meridian_transit(source, from_time, npoints=400):
     if not isinstance(source, SkyCoord):
         raise TypeError(
             'source must be a SkyCoord object'
-        )
-    if not isinstance(from_time, Time):
-        raise TypeError(
-            'from_time must be a Time object'
         )
     nenufar = Observer(
         name='NenuFAR',
@@ -551,3 +727,164 @@ def dispersion_delay(freq, dm):
     delay = 4140. * (dm/dm_ref) / ((freq/freq_ref)**2) * u.s
     return delay
 # ============================================================= #
+
+
+# ============================================================= #
+# ------------------------- _pos2ecef ------------------------- #
+# ============================================================= #
+def _pos2ecef(earthLocation=None):
+    """
+    """
+    return
+#     if earthLocation is None:
+#         earthLocation = nenufar_loc
+#     if not isinstance(earthLocation, EarthLocation):
+#         raise TypeError(
+#             'earthLocation should be a EarthLocation instance.'
+#         )
+#     lon = earthLocation.lon.rad
+#     lat = earthLocation.lat.rad, 
+#     alt = earthLocation.height.value
+#     gps_b = 6356752.31424518
+#     gps_a = 6378137
+#     e_squared = 6.69437999014e-3
+#     gps_n = gps_a / np.sqrt(1 - e_squared * np.sin(lat)**2)
+#     xyz = np.zeros((3))
+#     xyz[0] = (gps_n + alt) * np.cos(lat) * np.cos(lon)
+#     xyz[1] = (gps_n + alt) * np.cos(lat) * np.sin(lon)
+#     xyz[2] = (gps_b**2 / gps_a**2 * gps_n + alt) * np.sin(lat)
+#     return xyz
+# ============================================================= #
+
+
+# ============================================================= #
+# ------------------------- _ecef2enu ------------------------- #
+# ============================================================= #
+def _ecef2enu(ecef, earthLocation=None):
+    """
+    """
+    return
+#     xyz = np.array(xyz)
+#     if xyz.ndim > 1 and xyz.shape[1] != 3:
+#         raise ValueError("The expected shape of ECEF xyz array is (Npts, 3).")
+#     xyz_in = xyz
+
+#     if xyz_in.ndim == 1:
+#         xyz_in = xyz_in[np.newaxis, :]
+# ============================================================= #
+
+
+# gps_b = 6356752.31424518
+# gps_a = 6378137
+# e_squared = 6.69437999014e-3
+
+# def XYZ_from_LatLonAlt(latitude, longitude, altitude):
+#     """
+#     from pyuvdata
+#     Calculate ECEF x,y,z from lat/lon/alt values.
+
+#     Parameters
+#     ----------
+#     latitude :  ndarray or float
+#         latitude, numpy array (if Npts > 1) or value (if Npts = 1) in radians
+#     longitude :  ndarray or float
+#         longitude, numpy array (if Npts > 1) or value (if Npts = 1) in radians
+#     altitude :  ndarray or float
+#         altitude, numpy array (if Npts > 1) or value (if Npts = 1) in meters
+
+#     Returns
+#     -------
+#     xyz : ndarray of float
+#         numpy array, shape (Npts, 3), with ECEF x,y,z coordinates.
+
+#     """
+#     latitude = np.array(latitude)
+#     longitude = np.array(longitude)
+#     altitude = np.array(altitude)
+#     n_pts = latitude.size
+#     if longitude.size != n_pts:
+#         raise ValueError(
+#             "latitude, longitude and altitude must all have the same length"
+#         )
+#     if altitude.size != n_pts:
+#         raise ValueError(
+#             "latitude, longitude and altitude must all have the same length"
+#         )
+
+#     # see wikipedia geodetic_datum and Datum transformations of
+#     # GPS positions PDF in docs/references folder
+#     gps_n = gps_a / np.sqrt(1 - e_squared * np.sin(latitude) ** 2)
+#     xyz = np.zeros((n_pts, 3))
+#     xyz[:, 0] = (gps_n + altitude) * np.cos(latitude) * np.cos(longitude)
+#     xyz[:, 1] = (gps_n + altitude) * np.cos(latitude) * np.sin(longitude)
+#     xyz[:, 2] = (gps_b ** 2 / gps_a ** 2 * gps_n + altitude) * np.sin(latitude)
+
+#     xyz = np.squeeze(xyz)
+#     return xyz
+
+
+# def ENU_from_ECEF(xyz, latitude, longitude, altitude):
+#     """
+#     from pyuvdata
+#     Calculate local ENU (east, north, up) coordinates from ECEF coordinates.
+
+#     Parameters
+#     ----------
+#     xyz : ndarray of float
+#         numpy array, shape (Npts, 3), with ECEF x,y,z coordinates.
+#     latitude : float
+#         Latitude of center of ENU coordinates in radians.
+#     longitude : float
+#         Longitude of center of ENU coordinates in radians.
+#     altitude : float
+#         Altitude of center of ENU coordinates in radians.
+
+#     Returns
+#     -------
+#     ndarray of float
+#         numpy array, shape (Npts, 3), with local ENU coordinates
+
+#     """
+#     xyz = np.array(xyz)
+#     if xyz.ndim > 1 and xyz.shape[1] != 3:
+#         raise ValueError("The expected shape of ECEF xyz array is (Npts, 3).")
+
+#     xyz_in = xyz
+
+#     if xyz_in.ndim == 1:
+#         xyz_in = xyz_in[np.newaxis, :]
+
+#     # check that these are sensible ECEF values -- their magnitudes need to be
+#     # on the order of Earth's radius
+#     ecef_magnitudes = np.linalg.norm(xyz_in, axis=1)
+#     sensible_radius_range = (6.35e6, 6.39e6)
+#     if np.any(ecef_magnitudes <= sensible_radius_range[0]) or np.any(
+#         ecef_magnitudes >= sensible_radius_range[1]
+#     ):
+#         raise ValueError(
+#             "ECEF vector magnitudes must be on the order of the radius of the earth"
+#         )
+
+#     xyz_center = XYZ_from_LatLonAlt(latitude, longitude, altitude)
+
+#     xyz_use = np.zeros_like(xyz_in)
+#     xyz_use[:, 0] = xyz_in[:, 0] - xyz_center[0]
+#     xyz_use[:, 1] = xyz_in[:, 1] - xyz_center[1]
+#     xyz_use[:, 2] = xyz_in[:, 2] - xyz_center[2]
+
+#     enu = np.zeros_like(xyz_use)
+#     enu[:, 0] = -np.sin(longitude) * xyz_use[:, 0] + np.cos(longitude) * xyz_use[:, 1]
+#     enu[:, 1] = (
+#         -np.sin(latitude) * np.cos(longitude) * xyz_use[:, 0]
+#         - np.sin(latitude) * np.sin(longitude) * xyz_use[:, 1]
+#         + np.cos(latitude) * xyz_use[:, 2]
+#     )
+#     enu[:, 2] = (
+#         np.cos(latitude) * np.cos(longitude) * xyz_use[:, 0]
+#         + np.cos(latitude) * np.sin(longitude) * xyz_use[:, 1]
+#         + np.sin(latitude) * xyz_use[:, 2]
+#     )
+#     if len(xyz.shape) == 1:
+#         enu = np.squeeze(enu)
+
+#     return enu