+++ nenupy/observation/__init__.py
@@ -5,4 +5,5 @@
 from .tapdatabase import ObsDatabase
 from .sqldatabase import ParsetDataBase
 from .parset import Parset
+from .pointing import *